// frontend/types/animations.ts
/*!
 * This file is part of the Stooa codebase.
 *
 * (c) 2020 - present Runroom SL
 *
 * For the full copyright and license information, please view the LICENSE
 * file that was distributed with this source code.
 */

export interface Lottie {
  id: string;
  name: string;
  path: string;
  morph?: string;
  reverse?: boolean;
  assetsPath?: string;
}